refactor(testutil): organise network files
